backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
motor==3.6.0
openai>=1.99.9
python-multipart==0.0.12
